+++ examples/next-google-generative-ai/app/api/completion/route.ts
@@ -3,8 +3,7 @@ import { GoogleGenerativeAIStream, StreamingTextResponse } from 'ai';
 
 const genAI = new GoogleGenerativeAI(process.env.GOOGLE_API_KEY || '');
 
-// IMPORTANT! Set the runtime to edge
-export const runtime = 'edge';
+export const dynamic = 'force-dynamic';
 
 export async function POST(req: Request) {
   // Extract the `prompt` from the body of the request