+++ examples/next-huggingface/app/api/completion/route.ts
@@ -4,8 +4,7 @@ import { HuggingFaceStream, StreamingTextResponse } from 'ai';
 // Create a new Hugging Face Inference instance
 const Hf = new HfInference(process.env.HUGGINGFACE_API_KEY);
 
-// IMPORTANT! Set the runtime to edge
-export const runtime = 'edge';
+export const dynamic = 'force-dynamic';
 
 export async function POST(req: Request) {
   // Extract the `prompt` from the body of the request