+++ examples/next-openai-pages/pages/api/chat.ts
@@ -1,11 +1,11 @@
-import { OpenAI } from '@ai-sdk/openai';
+import { createOpenAI } from '@ai-sdk/openai';
 import { StreamingTextResponse, experimental_streamText } from 'ai';
 
-const openai = new OpenAI({
+const openai = createOpenAI({
   apiKey: process.env.OPENAI_API_KEY ?? '',
 });
 
-export const runtime = 'edge';
+export const dynamic = 'force-dynamic';
 
 // IMPORTANT: The Next.js Pages Router's API Routes do not support streaming responses.
 // see https://sdk.vercel.ai/docs/guides/frameworks/nextjs-pages
@@ -13,7 +13,7 @@ export default async function handler(req: Request, res: Response) {
   const { messages } = await req.json();
 
   const result = await experimental_streamText({
-    model: openai.chat('gpt-4-turbo-preview'),
+    model: openai('gpt-4-turbo-preview'),
     messages,
   });