+++ examples/next-openai/app/api/chat-with-vision/route.ts
@@ -6,8 +6,7 @@ const openai = new OpenAI({
   apiKey: process.env.OPENAI_API_KEY || '',
 });
 
-// IMPORTANT! Set the runtime to edge
-export const runtime = 'edge';
+export const dynamic = 'force-dynamic';
 
 export async function POST(req: Request) {
   // 'data' contains the additional data that you have sent: