+++ packages/core/core/generate-text/stream-text.test.ts
@@ -4,6 +4,7 @@ import { convertArrayToReadableStream } from '../test/convert-array-to-readable-
 import { convertAsyncIterableToArray } from '../test/convert-async-iterable-to-array';
 import { convertReadableStreamToArray } from '../test/convert-readable-stream-to-array';
 import { MockLanguageModelV1 } from '../test/mock-language-model-v1';
+import { createMockServerResponse } from '../test/mock-server-response';
 import { experimental_streamText } from './stream-text';
 
 describe('result.textStream', () => {
@@ -24,6 +25,7 @@ describe('result.textStream', () => {
               {
                 type: 'finish',
                 finishReason: 'stop',
+                logprobs: undefined,
                 usage: { completionTokens: 10, promptTokens: 3 },
               },
             ]),
@@ -59,6 +61,7 @@ describe('result.fullStream', () => {
               {
                 type: 'finish',
                 finishReason: 'stop',
+                logprobs: undefined,
                 usage: { completionTokens: 10, promptTokens: 3 },
               },
             ]),
@@ -78,6 +81,7 @@ describe('result.fullStream', () => {
         {
           type: 'finish',
           finishReason: 'stop',
+          logprobs: undefined,
           usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
         },
       ],
@@ -121,6 +125,7 @@ describe('result.fullStream', () => {
               {
                 type: 'finish',
                 finishReason: 'stop',
+                logprobs: undefined,
                 usage: { completionTokens: 10, promptTokens: 3 },
               },
             ]),
@@ -148,6 +153,7 @@ describe('result.fullStream', () => {
         {
           type: 'finish',
           finishReason: 'stop',
+          logprobs: undefined,
           usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
         },
       ],
@@ -191,6 +197,7 @@ describe('result.fullStream', () => {
               {
                 type: 'finish',
                 finishReason: 'stop',
+                logprobs: undefined,
                 usage: { completionTokens: 10, promptTokens: 3 },
               },
             ]),
@@ -226,6 +233,7 @@ describe('result.fullStream', () => {
         {
           type: 'finish',
           finishReason: 'stop',
+          logprobs: undefined,
           usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
         },
       ],
@@ -257,6 +265,139 @@ describe('result.toAIStream', () => {
       ),
       ['0:"Hello"\n', '0:", "\n', '0:"world!"\n'],
     );
+  });
+});
+
+describe('result.pipeAIStreamToResponse', async () => {
+  it('should write data stream parts to a Node.js response-like object', async () => {
+    const mockResponse = createMockServerResponse();
+
+    const result = await experimental_streamText({
+      model: new MockLanguageModelV1({
+        doStream: async () => {
+          return {
+            stream: convertArrayToReadableStream([
+              { type: 'text-delta', textDelta: 'Hello' },
+              { type: 'text-delta', textDelta: ', ' },
+              { type: 'text-delta', textDelta: 'world!' },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          };
+        },
+      }),
+      prompt: 'test-input',
+    });
+
+    result.pipeAIStreamToResponse(mockResponse);
+
+    // Wait for the stream to finish writing to the mock response
+    await new Promise(resolve => {
+      const checkIfEnded = () => {
+        if (mockResponse.ended) {
+          resolve(undefined);
+        } else {
+          setImmediate(checkIfEnded);
+        }
+      };
+      checkIfEnded();
+    });
+
+    const decoder = new TextDecoder();
+
+    assert.strictEqual(mockResponse.statusCode, 200);
+    assert.deepStrictEqual(mockResponse.headers, {
+      'Content-Type': 'text/plain; charset=utf-8',
+    });
+    assert.deepStrictEqual(
+      mockResponse.writtenChunks.map(chunk => decoder.decode(chunk)),
+      ['0:"Hello"\n', '0:", "\n', '0:"world!"\n'],
+    );
+  });
+});
+
+describe('result.pipeTextStreamToResponse', async () => {
+  it('should write text deltas to a Node.js response-like object', async () => {
+    const mockResponse = createMockServerResponse();
+
+    const result = await experimental_streamText({
+      model: new MockLanguageModelV1({
+        doStream: async () => {
+          return {
+            stream: convertArrayToReadableStream([
+              { type: 'text-delta', textDelta: 'Hello' },
+              { type: 'text-delta', textDelta: ', ' },
+              { type: 'text-delta', textDelta: 'world!' },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          };
+        },
+      }),
+      prompt: 'test-input',
+    });
+
+    result.pipeTextStreamToResponse(mockResponse);
+
+    // Wait for the stream to finish writing to the mock response
+    await new Promise(resolve => {
+      const checkIfEnded = () => {
+        if (mockResponse.ended) {
+          resolve(undefined);
+        } else {
+          setImmediate(checkIfEnded);
+        }
+      };
+      checkIfEnded();
+    });
+
+    const decoder = new TextDecoder();
+
+    assert.strictEqual(mockResponse.statusCode, 200);
+    assert.deepStrictEqual(mockResponse.headers, {
+      'Content-Type': 'text/plain; charset=utf-8',
+    });
+    assert.deepStrictEqual(
+      mockResponse.writtenChunks.map(chunk => decoder.decode(chunk)),
+      ['Hello', ', ', 'world!'],
+    );
+  });
+});
+
+describe('result.toAIStreamResponse', () => {
+  it('should create a Response with a stream data stream', async () => {
+    const result = await experimental_streamText({
+      model: new MockLanguageModelV1({
+        doStream: async ({ prompt, mode }) => {
+          return {
+            stream: convertArrayToReadableStream([
+              { type: 'text-delta', textDelta: 'Hello' },
+              { type: 'text-delta', textDelta: ', ' },
+              { type: 'text-delta', textDelta: 'world!' },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          };
+        },
+      }),
+      prompt: 'test-input',
+    });
+
+    const response = result.toAIStreamResponse();
+
+    assert.strictEqual(response.status, 200);
+    assert.strictEqual(
+      response.headers.get('Content-Type'),
+      'text/plain; charset=utf-8',
+    );
+
+    // Read the chunks into an array
+    const reader = response.body!.getReader();
+    const chunks = [];
+    while (true) {
+      const { value, done } = await reader.read();
+      if (done) break;
+      chunks.push(new TextDecoder().decode(value));
+    }
+
+    assert.deepStrictEqual(chunks, ['0:"Hello"\n', '0:", "\n', '0:"world!"\n']);
   });
 });