+++ packages/mistral/src/index.ts
@@ -1,1 +1,2 @@
 export * from './mistral-facade';
+export * from './mistral-provider';