+++ packages/provider/src/errors/index.ts
@@ -1,4 +1,5 @@
 export * from './api-call-error';
+export * from './empty-response-body-error';
 export * from './invalid-argument-error';
 export * from './invalid-data-content-error';
 export * from './invalid-prompt-error';
@@ -7,7 +8,6 @@ export * from './invalid-tool-arguments-error';
 export * from './json-parse-error';
 export * from './load-api-key-error';
 export * from './no-object-generated-error';
-export * from './no-response-body-error';
 export * from './no-such-tool-error';
 export * from './retry-error';
 export * from './tool-call-parse-error';