+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @ai-sdk/openai
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [56ef84a]
+  - @ai-sdk/provider-utils@0.0.4
+
+## 0.0.7
+
+### Patch Changes
+
+- 0833e19: Allow optional content to support Fireworks function calling.
+
+## 0.0.6
+
+### Patch Changes
+
+- d6431ae: ai/core: add logprobs support (thanks @SamStenner for the contribution)
+- 25f3350: ai/core: add support for getting raw response headers.
+- Updated dependencies [d6431ae]
+- Updated dependencies [25f3350]
+  - @ai-sdk/provider@0.0.2
+  - @ai-sdk/provider-utils@0.0.3
+
+## 0.0.5
+
+### Patch Changes
+
+- eb150a6: ai/core: remove scaling of setting values (breaking change). If you were using the temperature, frequency penalty, or presence penalty settings, you need to update the providers and adjust the setting values.
+- Updated dependencies [eb150a6]
+  - @ai-sdk/provider-utils@0.0.2
+  - @ai-sdk/provider@0.0.1
+
+## 0.0.4
+
+### Patch Changes
+
+- c6fc35b: Add custom header and OpenAI project support.
+
+## 0.0.3
+
+### Patch Changes
+
+- ab60b18: Simplified model construction by directly calling provider functions. Add create... functions to create provider instances.
+
 ## 0.0.2
 
 ### Patch Changes