+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/provider-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- 56ef84a: ai/core: fix abort handling in transformation stream
+
+## 0.0.3
+
+### Patch Changes
+
+- 25f3350: ai/core: add support for getting raw response headers.
+- Updated dependencies [d6431ae]
+- Updated dependencies [25f3350]
+  - @ai-sdk/provider@0.0.2
+
+## 0.0.2
+
+### Patch Changes
+
+- eb150a6: ai/core: remove scaling of setting values (breaking change). If you were using the temperature, frequency penalty, or presence penalty settings, you need to update the providers and adjust the setting values.
+- Updated dependencies [eb150a6]
+  - @ai-sdk/provider@0.0.1
+
 ## 0.0.1
 
 ### Patch Changes